% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_CNV.R
\name{feature_CNV}
\alias{feature_CNV}
\title{Get CNV}
\usage{
feature_CNV(path, bam)
}
\arguments{
\item{path, bam}{bamfile which you want to read}
}
\value{
CNV
}
\description{
A function, is used to get CNV.
}
