% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_NOM.R
\name{feature_NOM}
\alias{feature_NOM}
\title{Get NOM}
\usage{
feature_NOM(bamfile)
}
\arguments{
\item{bamfile}{bamfile which you want to read}
}
\value{
NOM
}
\description{
A function, is used to get NOM.
}
