% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_OJM.R
\name{feature_OJM}
\alias{feature_OJM}
\title{Get OJM}
\usage{
feature_OJM(bamfile)
}
\arguments{
\item{bamfile}{bamfile which you want to read}
}
\value{
OJM
}
\description{
A function, is used to get OJM.
}
