% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_OLR.R
\name{feature_OLR}
\alias{feature_OLR}
\title{Get OLR}
\usage{
feature_OLR(bamfile)
}
\arguments{
\item{bamfile}{bamfile which you want to read}
}
\value{
OLR
}
\description{
A function, is used to get OLR.
}
