% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_SBM.R
\name{feature_SBM}
\alias{feature_SBM}
\title{Get SBM}
\usage{
feature_SBM(bamfile)
}
\arguments{
\item{bamfile}{bamfile which you want to read}
}
\value{
SBM
}
\description{
A function, is used to get SBM.
}
